add newView Login

+++ app/src/main/java/com/example/salestapapp/supplier/data/database/dao/SuppliersDao.kt
@@ -23,6 +23,6 @@ interface SuppliersDao {
     suspend fun deleteSupplierByID(supplierID: Int)
 
     @Update
-    suspend fun editSupplier(suppliers: SupplierEntity):Long
+    suspend fun editSupplier(suppliers: SupplierEntity):Int
 
 }
